+++ modules/struct/tree/walk.c
@@ -1,8 +1,9 @@
 
+#include <stdint.h>
 #include <string.h>
-#include "tcl.h"
-#include <t.h>
-#include <util.h>
+#include <tcl.h>
+#include "t.h"
+#include "util.h"
 
 /* .................................................. */
 
@@ -237,7 +238,7 @@ t_walk_invokecmd (Tcl_Interp* interp, TN* n, Tcl_Obj* dummy0,
 		  Tcl_Obj* action)
 {
     int	      res;
-    int	      cc = (int)       dummy0;
+    int	      cc = (intptr_t)  dummy0;
     Tcl_Obj** ev = (Tcl_Obj**) dummy1; /* cc+3 elements */
 
     ev [cc]   = dummy2;	   /* Tree */
